Delete ref

+++ public/application/services/budget.service.js
@@ -100,37 +100,10 @@ app.service("budgetService", [
             var sheets = ctx.workbook.worksheets;
             sheets.load("items");
             var ids = _.map(sheets.items, function(sheet) { return sheet.id });
-            _.forEach(ids, function (id) {
-              ws = ctx.workbook.worksheets.getItem(id);
-              ws.delete();
-            });
-           
-            return ctx.sync()
-            .then(function (response) {
-              next(null, data);  
-            }).catch(function (err) {
-              next(null, data);
-            });
-          }).catch(function (err) {
-            next(err);
-          });
-      });
-    }
-
-
-    var deleteWorkSheets = function (data, next) {
-      Excel.run(function (ctx) {
-        var sheets = ctx.workbook.worksheets;
-        sheets.load("items");
-        var count = ctx.workbook.worksheets.getCount();
-        return ctx.sync()
-          .then(function(response) {
-            var sheets = ctx.workbook.worksheets;
-            sheets.load("items");
-            var ids = _.map(sheets.items, function(sheet) { return sheet.id });
-            _.forEach(ids, function (id) {
-              ws = ctx.workbook.worksheets.getItem(id);
-              ws.delete();
+            _.forEach(ids, function (id, key) {
+              var ws = ctx.workbook.worksheets.getItem(id);
+              if (key < ids.length - 1) 
+                ws.delete();
             });
            
             return ctx.sync()
